extract check_key into utils
